+++ source/UberCore/Networking/Request.swift
@@ -105,10 +105,14 @@ public class Request {
      */
     private func addHeaders() {
         urlRequest.setValue("gzip, deflate", forHTTPHeaderField: "Accept-Encoding")
+
+        if let versionNumber = Bundle(for: type(of: self)).object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String {
+            urlRequest.setValue("iOS Rides SDK v\(versionNumber)", forHTTPHeaderField: "X-Uber-User-Agent")
+        }
         if let token = bearerToken {
-            urlRequest.setValue("Bearer \(token)", forHTTPHeaderField: HTTPHeader.Authorization.rawValue)
+            urlRequest.setValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
         } else if let token = serverToken {
-            urlRequest.setValue("Token \(token)", forHTTPHeaderField: HTTPHeader.Authorization.rawValue)
+            urlRequest.setValue("Token \(token)", forHTTPHeaderField: "Authorization")
         }
         if let headers = endpoint.headers {
             for (header,value) in headers {