+++ source/UberRides/RideRequestViewErrorType.swift
@@ -31,6 +31,8 @@ Possible errors that can occur in the RideRequestView.
 - networkError:               A network error occured
 - notSupported:               The functionality attempted is not supported on the current device
 - unknown:                    Unknown error occured.
+- Warning: The Ride Request Widget is deprecated, and will no longer work for new apps.
+Existing apps have until 05/31/2018 to migrate. See the Uber API Changelog for more details.
 */
 @objc public enum RideRequestViewErrorType: Int {
     case accessTokenExpired